backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
werkzeug==2.0.3
sqlalchemy==1.4.46
Flask-SQLAlchemy==2.5.1
pydantic>=1.10.0,<2.0.0
psycopg2-binary==2.9.9
PyJWT==2.1.0